refactoring_all_js_function
