+++ javascript/today.js
@@ -1,53 +1,5 @@
-function create_array() {
-    fetch(`http://127.0.0.1:8000/today_word_list/total/`, {
-        method:'GET',
-        headers:{
-            'Content-Type': 'application/text'
-        }})
-    .then((response) => response.text().then(function (myjsonarray) {
-            list_start_point = document.getElementsByClassName('word_today')[0]
-            let jsondatas = JSON.parse(myjsonarray)
-            for (let i = 0; i < jsondatas.length; i++){
-                let row = `<span class="e${i+1}" style = "width: 420px;
-                color: rgba(0, 0, 0, 1);
-                position: absolute;
-                top: ${5 + (i*34)}px;
-                left: 10px;
-                font-family: Roboto;
-                font-weight: Thin;
-                font-size: 20px;
-                opacity: 1;
-                text-align: left;">${jsondatas[i].word}</span>`
-                list_start_point.innerHTML += row;
-            }
-            for (let i = 0; i < jsondatas.length; i++){
-                let row = `<span class="m${i+1}" style = "width: 420px;
-                color: rgba(0, 0, 0, 1);
-                position: absolute;
-                top: ${5 + (i*34)}px;
-                left: 480px;
-                font-family: Roboto;
-                font-weight: Thin;
-                font-size: 20px;
-                opacity: 1;
-                text-align: left;">${jsondatas[i].meaning}</span>`
-                list_start_point.innerHTML += row;
-            }
-            for (let i = 0; i < jsondatas.length; i++){
-                let row = `<span class="sen${i+1}" style="width: 622px;
-                color: rgba(0, 0, 0, 1);
-                position: absolute;
-                top: ${5 + (i*34)}px;
-                left: 950px;
-                font-family: Roboto;
-                font-weight: Thin;
-                font-size: 20px;
-                opacity: 1;
-                text-align: left;">${jsondatas[i].sentence}</span>`
-                list_start_point.innerHTML += row;
-            }
-        }));
-}
+import get_today from './all_share.js';
+import create_array from './partial_share.js';
 
 function go_total() {
     fetch(`http://127.0.0.1:8000/today_to_total/total/`, {
@@ -73,18 +25,13 @@ function go_done() {
         })
 }
 
-
-function get_today() {
-    var dt = new Date();
-    const year = dt.getFullYear();
-    const month = dt.getMonth();
-    const date = dt.getDate();
-    document.getElementsByClassName('current_date')[0].innerHTML = `${year}년 ${month >= 10 ? month : '0' + (month+1)}월 ${date >= 10 ? date : '0' + date}일`;
-}
-
 function init_loader() {
-    get_today();
+    get_today(document);
     console.log("time yes");
-    create_array();
+    create_array('today_word_list', 'word_today', document);
     console.log("array yes");
 }
+
+document.addEventListener("DOMContentLoaded", () => {
+    init_loader();
+})
