+++ templete/first_page.html
@@ -4,9 +4,9 @@
     <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet"/>
     <link href="/templete/first_page.css" rel="stylesheet"/>
     <title>Document</title>
-    <script src="/javascript/move.js"></script>
+    <script type="module" src="/javascript/move.js"></script>
 </head>
-<body onload="get_today()">
+<body>
 <div class="first_page">
     <form method="GET">
         <button class="web_name" onclick="reload_alert()">편입영어</button>
