+++ templete/today.html
@@ -1,14 +1,12 @@
 <!DOCTYPE html>
 <html>
-
 <head>
     <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet" />
     <link href="/templete/today.css" rel="stylesheet" />
     <title>Document</title>
+    <script type="module" src="/javascript/today.js"></script>
 </head>
-<script src="/javascript/today.js"></script>
-
-<body onload="init_loader()">
+<body>
     <div class="today">
         <button class="web_name" onclick="location.href='first_page'">편입영어</button>
         <span class="current_date"></span>
