+++ templete/total.html
@@ -5,10 +5,9 @@
     <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet"/>
     <link href="/templete/total.css" rel="stylesheet"/>
     <title>Document</title>
+    <script type="module" src="/javascript/total.js"></script>
 </head>
-<script src="/javascript/total.js"></script>
-
-<body onload="init_loader()">
+<body>
     <div class="total">
         <button class="web_name" onclick="location.href='first_page'">편입영어</button>
         <span class="current_date"></span>
