+++ templete/total_review.html
@@ -4,9 +4,9 @@
     <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet" />
     <link href="/templete/total_review.css" rel="stylesheet" />
     <title>Document</title>
-    <script src="/javascript/total_review.js"></script>
+    <script type="module" src="/javascript/total_review.js"></script>
 </head>
-<body onload="first_load()">
+<body>
 <div class="total_review">
     <form method="GET">
         <button class="web_name" onclick="reload_alert()">편입영어</button>
